ml/requirements.txt
numpy
matplotlib
scikit-learn
imageio
imageio-ffmpeg
torch
torchvision
pandas
seaborn
nltk

transformers
datasets

python-gflags
loguru

tensorflow_datasets


jax